+++ src/Discord/Parts/User/Member.php
@@ -199,49 +199,6 @@ class Member extends Part implements Stringable
     }
 
     /**
-     * Modifies the current member (no validation).
-     *
-     * @link https://discord.com/developers/docs/resources/guild#modify-current-member-json-params
-     *
-     * @since 10.19.0
-     *
-     * @param array        $params           The parameters to modify.
-     * @param ?string|null $params['nick']   Value to set user's nickname to.
-     * @param ?string|null $params['banner'] Data URI base64 encoded banner image.
-     * @param ?string|null $params['avatar'] Data URL base64 encoded avatar image.
-     * @param ?string|null $params['bio']    Guild member bio.
-     * @param string|null  $reason           Reason for Audit Log.
-     *
-     * @throws NoPermissionsException Member is not the current user.
-     *
-     * @return PromiseInterface<self>
-     */
-    public function modifyCurrentMember(array $params, ?string $reason = null): PromiseInterface
-    {
-        if ($this->discord->id != $this->id) {
-            return reject(new NoPermissionsException('You can only modify the current member.'));
-        }
-
-        $allowed = ['nick', 'banner', 'avatar', 'bio'];
-        $params = array_filter(
-            $params,
-            fn ($key) => in_array($key, $allowed, true),
-            ARRAY_FILTER_USE_KEY
-        );
-
-        if (empty($params)) {
-            return reject(new \InvalidArgumentException('No valid parameters to modify.'));
-        }
-
-        $headers = [];
-        if (isset($reason)) {
-            $headers['X-Audit-Log-Reason'] = $reason;
-        }
-
-        return $this->http->patch(Endpoint::bind(Endpoint::GUILD_MEMBER_SELF, $this->guild_id), $params, $headers);
-    }
-
-    /**
      * Sets the nickname of the member.
      *
      * @param ?string|null $nick   The nickname of the member.
